+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Added move markers on click & drag #35.
 *   Added dual markers "Scale to fit" feature #66.
 *   Added keyboard movement/zoom of waveform display #11.
+*   Added marker colors #57.
 
 
 ## 0.8.16